ANDROID-13326 NestedScrollWebView refactor, fixes for detected issues and improvements
