+++ app/src/main/res/layout/activity_scrolling.xml
@@ -43,7 +43,8 @@
         <com.telefonica.nestedscrollwebview.NestedScrollWebView
             android:id="@+id/webView"
             android:layout_width="match_parent"
-            android:layout_height="match_parent" />
+            android:layout_height="match_parent"
+            app:blockNestedScrollingOnHorizontalScrolls="true"/>
 
     </androidx.swiperefreshlayout.widget.SwipeRefreshLayout>
