--- nestedscrollwebview/src/main/java/com/telefonica/nestedscrollwebview/helper/NestedScrollingView.kt
@@ -1,20 +1,0 @@
-package com.telefonica.nestedscrollwebview.helper
-
-import android.view.ViewGroup
-
-interface NestedScrollingView {
-    val view: ViewGroup
-    fun getScrollRange(): Int
-
-    /* Existing ViewGroup methods with protected visibility */
-    fun computeHorizontalScrollRange(): Int
-    fun computeHorizontalScrollExtent(): Int
-    fun computeVerticalScrollRange(): Int
-    fun computeVerticalScrollExtent(): Int
-    fun onOverScrolled(
-        scrollX: Int,
-        scrollY: Int,
-        clampedX: Boolean,
-        clampedY: Boolean
-    )
-}
