+++ nestedscrollwebview/src/main/res/values/attrs.xml
@@ -2,5 +2,6 @@
 <resources>
     <declare-styleable name="NestedScrollWebView">
         <attr name="coordinatorBottomMatchingEnabled" format="boolean" />
+        <attr name="blockNestedScrollingOnHorizontalScrolls" format="boolean" />
     </declare-styleable>
 </resources>
